[ADD] Saving objects paths

+++ gmath/random.go
@@ -6,14 +6,14 @@ func GetRandomInt(min int, max int) int {
 	return rand.Intn(max-min) + min
 }
 
-func GetRandomFloat(min float64, max float64) float64 {
-	return rand.Float64()*(max-min) + min
+func GetRandomFloat(min float32, max float32) float32 {
+	return rand.Float32()*(max-min) + min
 }
 
 func GetRandomBool() bool {
 	return rand.Intn(2) == 1
 }
 
-func GetRandomOnPoint(point float64, delta float64) float64 {
+func GetRandomOnPoint(point float32, delta float32) float32 {
 	return GetRandomFloat(point-delta, point+delta)
 }
